Discard input before reading from stdin
Try to read a total of 65k but with a timeout of 1ms. This effectively
discards any input which has accumulated while 'show' was running. (Yes,
this fails if there are more than 65k bytes pending, but I'd like to see
that actually happen.)

The purpose of this is to eliminate "lag" when a user presses a key
constantly and then releases it. Without discarding outstanding input,
we would be processing all the buffered keystrokes.